vivimagazine/configs.py
import os
from dotenv import load_dotenv

load_dotenv()


class Configs:
    SECRET_KEY = os.getenv('SECRET_KEY', 'batata')
    DATABASE_USER = os.getenv("DATABASE_USER")
    DATABASE_PASSWORD = os.getenv("DATABASE_PASSWORD")
    DATABASE_NAME = os.getenv("DATABASE_NAME")
    DATABASE_HOST = os.getenv("DATABASE_HOST")
    DATABASE_PORT = os.getenv("DATABASE_PORT")
