requirements.txt
psycopg2-binary
Django
Pillow
python-dotenv
djangorestframework